Change speedometer.speed package name to => speedometer
