+++ android/src/com/mobiperf/speedometer/AccountSelector.java
@@ -12,7 +12,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.mobiperf.speedometer.speed;
+package com.mobiperf.speedometer;
 
 import android.accounts.Account;
 import android.accounts.AccountManager;
