+++ android/src/com/mobiperf/speedometer/MeasurementResult.java
@@ -12,7 +12,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.mobiperf.speedometer.speed;
+package com.mobiperf.speedometer;
 
 import com.mobiperf.speedometer.measurements.DnsLookupTask;
 import com.mobiperf.speedometer.measurements.DnsLookupTask.DnsLookupDesc;
