+++ android/src/com/mobiperf/speedometer/ResultsConsoleActivity.java
@@ -12,7 +12,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.mobiperf.speedometer.speed;
+package com.mobiperf.speedometer;
 
 import com.mobiperf.chart.CubicChart;
 import com.mobiperf.mobiperf.MobiperfActivity;
