+++ android/src/com/mobiperf/speedometer/WatchdogBootReceiver.java
@@ -13,7 +13,7 @@
  * limitations under the License.
  */
 
-package com.mobiperf.speedometer.speed;
+package com.mobiperf.speedometer;
 
 import com.mobiperf.mobiperf.R;
