+++ android/src/com/mobiperf/speedometer/measurements/UDPBurstTask.java
@@ -2,10 +2,10 @@
 
 package com.mobiperf.speedometer.measurements;
 
-import com.mobiperf.speedometer.speed.MeasurementDesc;
-import com.mobiperf.speedometer.speed.MeasurementError;
-import com.mobiperf.speedometer.speed.MeasurementResult;
-import com.mobiperf.speedometer.speed.MeasurementTask;
+import com.mobiperf.speedometer.MeasurementDesc;
+import com.mobiperf.speedometer.MeasurementError;
+import com.mobiperf.speedometer.MeasurementResult;
+import com.mobiperf.speedometer.MeasurementTask;
 import com.mobiperf.speedometer.util.PhoneUtils;
 
 import android.content.Context;
